+++ ls_Monobrow/core/changelog.lua
@@ -6,7 +6,5 @@ local _G = getfenv(0)
 
 -- Mine
 addon.CHANGELOG = [[
-- Added proper handling of Edit Mode layout copying, removal, an deletion.
-- Reduced the min width to 384px.
-- Fixed an issue where the monobar wouldn't hide itself while in a vehicle or during a pet battle.
+- Added options to fade in the monobar on entering combat and/or acquiring a target.
 ]]