+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # CHANGELOG
 
+## Version 120000.04
+
+- Added options to fade in the monobar on entering combat and/or acquiring a target.
+
 ## Version 120000.03
 
-- Added proper handling of Edit Mode layout copying, removal, an deletion.
+- Added proper handling of Edit Mode layout copying, removal, and deletion.
 - Reduced the min width to 384px.
 - Fixed an issue where the monobar wouldn't hide itself while in a vehicle or during a pet battle.